Предрелиз

+++ src/WebApi/Models/UserBase.cs
@@ -1,7 +1,11 @@
 ﻿namespace WebApi.Models
 {
+    /// <summary>
+    /// Родитель для всех классов - образов таблиц в БД
+    /// Определяет ключ таблицы
+    /// </summary>
     public abstract class UserBase
     {
-        public long id { get; set; }
+        public int id { get; set; }
     }
 }
